refactor: move Selenium tests to tests/selenium/
Names the directories after the tool that drives them, freeing
tests/integration/ for the Playwright suite. Pure rename plus the
shared create_test_device import path; the integration tag comes from
SeleniumTestCase, so discovery and 'invoke tests --tag integration'
are unaffected.
