+++ nautobot/core/tests/selenium/test_app_navbar.py
